server/requirements/test.txt
-r base.txt

behave==1.2.5
behave-django==1.0.0
parse==1.8.2
parse-type==0.4.2
six==1.11.0
jsondiff==1.1.1
